🔥 Only re-export langs, don't register
