+++ packages/all/index.js
@@ -1,2 +1,2 @@
+export * from './lang.js'
 export * from './langs.js'
-export * from './register.js'
